move custom rule file to test-util/
