+++ src/test-util/prefer-addition-shorthand-rule.ts
